requirements.txt
pytest~=8.3.5
selene == 2.0.0rc9

selenium
allure-pytest==2.14.0
allure-python-commons~=2.14.0
python-dotenv
